[v 1.0.0 ] [refactor: AlignedCollectionViewFlowLayout 분기 추가로 인한 리팩토링 ]
